Added 4-byte fill field before 64-bit p.rtime to fix alignment problems due to new cloud fields taking up (1 + 3 * MAXLEV) 4-byte words, an odd number of words.  This fill field was there in the past, but commented out for rtpV201, which had p.rtime on an even 4-byte word boundary.
Did more tests, including using flag values at the ends of key
arrays, comparing output from the C and matlab versions of rtpdump,
and checking that IRCALCBIT and IROBSVBIT still work as expected.

A quick check of bumping MAXCHAN up to 5000 channels for both robs1
and rcalc suggests that this still doesn't work.