updated rtpsize to check if a fill value is needed before p.rtime. Note that it's OK if the struct size is one word bigger than than the struct span, it's the latter that is compared with the field name lists (FLISTs), with a warning if there is a mismatch.
further tests suggest that with a fill value (if needed), if rtpsize
does not give any warnings, and if the Fortran tests are OK, then we
don't really need the struct packing compiler options.  The Makefile
is updated to reflect this.

updated the README, but not the rtpspec docs, yet.